refactor: исправил замечания
