+++ main-service/src/main/java/ru/practicum/ewm/main/service/category/dto/CategoryDto.java
@@ -3,7 +3,6 @@ package ru.practicum.ewm.main.service.category.dto;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 import org.hibernate.validator.constraints.Length;
-import ru.practicum.ewm.main.service.category.model.Category;
 
 import javax.validation.constraints.NotBlank;
 
@@ -20,12 +19,5 @@ public class CategoryDto {
     @NotBlank
     @Length(max = 50)
     String name;
-
-    public static CategoryDto mapToCategoryDto(Category category) {
-        return CategoryDto.builder()
-                .id(category.getId())
-                .name(category.getName())
-                .build();
-    }
 
 }
