+++ main-service/src/main/java/ru/practicum/ewm/main/service/category/mapper/CategoryMapper.java
@@ -1,0 +1,15 @@
+package ru.practicum.ewm.main.service.category.mapper;
+
+import org.mapstruct.Mapper;
+import org.mapstruct.factory.Mappers;
+import ru.practicum.ewm.main.service.category.dto.CategoryDto;
+import ru.practicum.ewm.main.service.category.model.Category;
+
+@Mapper
+public interface CategoryMapper {
+
+    CategoryMapper CATEGORY_MAPPER = Mappers.getMapper(CategoryMapper.class);
+
+    CategoryDto mapToCategoryDto(Category category);
+
+}
