+++ main-service/src/main/java/ru/practicum/ewm/main/service/event/dto/EventDto.java
@@ -2,12 +2,10 @@ package ru.practicum.ewm.main.service.event.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
-import ru.practicum.ewm.main.service.category.dto.CategoryDto;
-import ru.practicum.ewm.main.service.event.model.Event;
-import ru.practicum.ewm.main.service.event.model.EventState;
-import ru.practicum.ewm.main.service.rate.dto.RateDto;
-import ru.practicum.ewm.main.service.user.dto.UserSimpleDto;
 import lombok.experimental.FieldDefaults;
+import ru.practicum.ewm.main.service.category.dto.CategoryDto;
+import ru.practicum.ewm.main.service.event.model.EventState;
+import ru.practicum.ewm.main.service.user.dto.UserSimpleDto;
 
 import java.time.LocalDateTime;
 
@@ -39,31 +37,8 @@ public class EventDto {
     EventState state;
     String title;
     long views;
-    RateDto rate;
-
-
-
-    public static EventDto mapToEventDto(Event event) {
-        return EventDto.builder()
-                .id(event.getId())
-                .title(event.getTitle())
-                .annotation(event.getAnnotation())
-                .description(event.getDescription())
-                .category(CategoryDto.mapToCategoryDto(event.getCategory()))
-                .initiator(UserSimpleDto.mapToUserSimpleDto(event.getInitiator()))
-                .eventDate(event.getEventDate())
-                .createdOn(event.getCreatedOn())
-                .publishedOn(event.getPublishedOn())
-                .state(event.getState())
-                .location(LocationDto.mapToLocationDto(event.getLocation()))
-                .participantLimit(event.getParticipantLimit())
-                .requestModeration(event.isRequestModeration())
-                .confirmedRequests(event.getConfirmedRequests())
-                .views(event.getViews())
-                .paid(event.isPaid())
-                .rate(event.getRate())
-                .build();
-    }
-
+    long likes;
+    long dislikes;
+    long rating;
 
 }
