+++ main-service/src/main/java/ru/practicum/ewm/main/service/event/model/Event.java
@@ -3,8 +3,6 @@ package ru.practicum.ewm.main.service.event.model;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 import ru.practicum.ewm.main.service.category.model.Category;
-import ru.practicum.ewm.main.service.event.dto.EventCreateDto;
-import ru.practicum.ewm.main.service.rate.dto.RateDto;
 import ru.practicum.ewm.main.service.user.model.User;
 
 import javax.persistence.*;
@@ -45,7 +43,7 @@ public class Event implements Serializable {
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
     EventState state;
-    @OneToOne(fetch = FetchType.EAGER)
+    @OneToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
     @JoinColumn(name = "location_id", nullable = false)
     Location location;
     @ManyToOne(fetch = FetchType.EAGER)
@@ -59,23 +57,8 @@ public class Event implements Serializable {
     @Transient
     int confirmedRequests;
     @Transient
-    RateDto rate;
-
-    public long getRating() {
-        return rate.getLikes() - rate.getDislikes();
-    }
-
-    public static Event mapToEvent(EventCreateDto dto) {
-        return Event.builder()
-                .title(dto.getTitle())
-                .description(dto.getDescription())
-                .annotation(dto.getAnnotation())
-                .eventDate(dto.getEventDate())
-                .location(Location.mapToLocation(dto.getLocation()))
-                .participantLimit(dto.getParticipantLimit())
-                .requestModeration(dto.isRequestModeration())
-                .paid(dto.isPaid())
-                .build();
-    }
+    long likes;
+    @Transient
+    long dislikes;
 
 }
