+++ main-service/src/main/java/ru/practicum/ewm/main/service/event/service/EventService.java
@@ -11,22 +11,18 @@ public interface EventService {
 
     EventDto save(int userId, EventCreateDto dto);
 
-    Event getByInitiatorAndId(int userId, int eventId);
-
     EventDto updateEvent(int userId, int eventId, EventUpdateDto updateRequest);
 
     List<EventDto> getAllEventsAdmin(GetEventsAdminDto params);
 
     EventDto updateEventAdmin(int eventId, EventUpdateDto dto);
 
-    List<EventSimpleDto> getAllPublishedEvents(GetEventsDto params);
-
     Event getEventById(int id);
 
     EventDto getPublishedEventById(int eventId);
 
-    List<Event> findAllById(List<Integer> events);
+    List<EventSimpleDto> getAllPublishedEvents(GetEventsDto params);
 
-    List<Event> setRates(List<Event> events);
+    Event getByInitiatorAndId(int userId, int eventId);
 
 }
