+++ main-service/src/main/java/ru/practicum/ewm/main/service/rate/user_rate/service/UserRateService.java
@@ -1,0 +1,11 @@
+package ru.practicum.ewm.main.service.rate.user_rate.service;
+
+public interface UserRateService {
+
+    void rateUser(boolean isLike, int fanId, int userId);
+
+    void updateUserRate(boolean isLike, int fanId, int userId);
+
+    void deleteUserRate(int fanId, int userId);
+
+}
