+++ main-service/src/main/java/ru/practicum/ewm/main/service/user/dto/UserDto.java
@@ -4,8 +4,6 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 import org.hibernate.validator.constraints.Length;
-import ru.practicum.ewm.main.service.rate.dto.RateDto;
-import ru.practicum.ewm.main.service.user.model.User;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
@@ -28,17 +26,9 @@ public class UserDto {
     @NotBlank
     @Length(min = 6, max = 254)
     String email;
-    RateDto rate;
+    long likes;
+    long dislikes;
+    long rating;
     Long privateRating;
-
-    public static UserDto mapToUserDto(User user) {
-        return UserDto.builder()
-                .id(user.getId())
-                .name(user.getName())
-                .email(user.getEmail())
-                .rate(user.getRate())
-                .privateRating(user.getPrivateRating())
-                .build();
-    }
 
 }
