+++ main-service/src/main/java/ru/practicum/ewm/main/service/user/dto/UserSimpleDto.java
@@ -2,8 +2,6 @@ package ru.practicum.ewm.main.service.user.dto;
 
 import lombok.*;
 import lombok.experimental.FieldDefaults;
-import ru.practicum.ewm.main.service.rate.dto.RateDto;
-import ru.practicum.ewm.main.service.user.model.User;
 
 @Setter
 @Getter
@@ -14,14 +12,8 @@ public class UserSimpleDto {
 
     int id;
     String name;
-    RateDto rate;
-
-    public static UserSimpleDto mapToUserSimpleDto(User user) {
-        return UserSimpleDto.builder()
-                .id(user.getId())
-                .name(user.getName())
-                .rate(user.getRate())
-                .build();
-    }
+    long likes;
+    long dislikes;
+    long rating;
 
 }
