+++ main-service/src/main/java/ru/practicum/ewm/main/service/user/service/UserServiceImpl.java
@@ -9,13 +9,14 @@ import org.springframework.stereotype.Service;
 import ru.practicum.ewm.main.service.event.dto.SortParam;
 import ru.practicum.ewm.main.service.exception.AlreadyExistsException;
 import ru.practicum.ewm.main.service.exception.NotFoundException;
-import ru.practicum.ewm.main.service.rate.dto.RateDto;
+import ru.practicum.ewm.main.service.rate.dto.Rate;
 import ru.practicum.ewm.main.service.rate.event_rate.EventRateRepository;
 import ru.practicum.ewm.main.service.rate.user_rate.UserRateRepository;
 import ru.practicum.ewm.main.service.user.UserRepository;
 import ru.practicum.ewm.main.service.user.dto.UserDto;
 import ru.practicum.ewm.main.service.user.model.User;
 
+import javax.transaction.Transactional;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
@@ -23,10 +24,12 @@ import java.util.stream.Collectors;
 
 import static ru.practicum.ewm.main.service.event.dto.SortParam.PRIVATE_RATING;
 import static ru.practicum.ewm.main.service.event.dto.SortParam.RATING;
-import static ru.practicum.ewm.main.service.rate.dto.RateDto.mapDbResultToIdAndRateDtoMap;
+import static ru.practicum.ewm.main.service.rate.dto.Rate.mapDbResultToIdAndRateDtoMap;
+import static ru.practicum.ewm.main.service.user.mapper.UserMapper.USER_MAPPER;
 
 @Slf4j
 @Service
+@Transactional
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
 
@@ -39,9 +42,8 @@ public class UserServiceImpl implements UserService {
         validateEmail(dto.getEmail());
         User user = User.mapToUser(dto);
         user = repo.save(user);
-        user.setRate(new RateDto());
         log.info("Пользователь зарегистрирован! {}", user);
-        return UserDto.mapToUserDto(user);
+        return USER_MAPPER.mapToUserDto(user);
     }
 
     @Override
@@ -54,13 +56,13 @@ public class UserServiceImpl implements UserService {
     @Override
     public List<UserDto> getAll(List<Integer> ids, int from, int size, SortParam sortParam) {
         List<User> users;
-        Comparator<User> comparator;
+        Comparator<UserDto> comparator;
         if (RATING.equals(sortParam)) {
-            comparator = Comparator.comparing(User::getRating).reversed();
+            comparator = Comparator.comparing(UserDto::getRating).reversed();
         } else if (PRIVATE_RATING.equals(sortParam)) {
-            comparator = Comparator.comparing(User::getPrivateRating).reversed();
+            comparator = Comparator.comparing(UserDto::getPrivateRating).reversed();
         } else {
-            comparator = Comparator.comparingInt(User::getId);
+            comparator = Comparator.comparingInt(UserDto::getId);
         }
         if (ids != null && !ids.isEmpty()) {
             users = repo.findAllByIdIn(ids);
@@ -74,8 +76,8 @@ public class UserServiceImpl implements UserService {
             users = repo.findAll(pageable).getContent();
         }
         return setRates(users, true).stream()
+                .map(USER_MAPPER::mapToUserDto)
                 .sorted(comparator)
-                .map(UserDto::mapToUserDto)
                 .collect(Collectors.toList());
     }
 
@@ -93,13 +95,16 @@ public class UserServiceImpl implements UserService {
     }
 
     private List<User> setRates(List<User> users, boolean isPrivate) {
-        Map<Integer, RateDto> rates = mapDbResultToIdAndRateDtoMap(userRateRepo.getRatesByUserIn(users));
-        Map<Integer, RateDto> privateRates = isPrivate ?
+        Map<Integer, Rate> rates = mapDbResultToIdAndRateDtoMap(userRateRepo.getRatesByUserIn(users));
+        Map<Integer, Rate> privateRates = isPrivate ?
                 mapDbResultToIdAndRateDtoMap(eventRateRepo.getPrivateRatesByUserIn(users))
                 : Map.of();
         users.forEach(user -> {
-            user.setRate(rates.getOrDefault(user.getId(), new RateDto()));
-            user.setPrivateRating(privateRates.getOrDefault(user.getId(), new RateDto()).getRating());
+            Rate rate = rates.getOrDefault(user.getId(), new Rate());
+            user.setLikes(rate.getLikes());
+            user.setDislikes(rate.getDislikes());
+            Rate privateRate = privateRates.getOrDefault(user.getId(), new Rate());
+            user.setPrivateRating(privateRate.getLikes() - privateRate.getDislikes());
         });
         return users;
     }
