+++ postman/feature.json
@@ -242,7 +242,9 @@
 									"pm.expect(target).to.have.property('publishedOn');\r",
 									"pm.expect(target).to.have.property('location');\r",
 									"pm.expect(target).to.have.property('requestModeration');\r",
-									"pm.expect(target).to.have.property('rate');\r",
+									"pm.expect(target).to.have.property('rating');\r",
+									"pm.expect(target).to.have.property('likes');\r",
+									"pm.expect(target).to.have.property('dislikes');\r",
 									"});\r",
 									"\r",
 									"pm.test(\"Данные в ответе должны соответствовать данным в запросе\", function () {\r",
@@ -506,7 +508,9 @@
 									"pm.expect(target).to.have.property('publishedOn');\r",
 									"pm.expect(target).to.have.property('location');\r",
 									"pm.expect(target).to.have.property('requestModeration');\r",
-									"pm.expect(target).to.have.property('rate');\r",
+									"pm.expect(target).to.have.property('rating');\r",
+									"pm.expect(target).to.have.property('likes');\r",
+									"pm.expect(target).to.have.property('dislikes');\r",
 									"});\r",
 									"\r",
 									"pm.test(\"Данные в ответе должны соответствовать данным в запросе\", function () {\r",
@@ -517,8 +521,8 @@
 									"    pm.expect(source.description).equal(target.description, 'Описание события должно соответствовать описанию события с указанным идентификатором');\r",
 									"    pm.expect(source.title).equal(target.title, 'Название события должно соответствовать названию события с указанным идентификатором');\r",
 									"    pm.expect(source.participantLimit.toString()).equal(target.participantLimit.toString(), 'Лимит участников события должен соответствовать лимиту участников события с указанным идентификатором');\r",
-									"    pm.expect(rate.likes.toString()).equal(target.rate.likes.toString(), 'Количество лайков должено соответствовать количеству лайков события с указанным идентификатором');\r",
-									"    pm.expect(rate.dislikes.toString()).equal(target.rate.dislikes.toString(), 'Количество дизлайков должено соответствовать количеству дизлайков события с указанным идентификатором');\r",
+									"    pm.expect(rate.likes.toString()).equal(target.likes.toString(), 'Количество лайков должено соответствовать количеству лайков события с указанным идентификатором');\r",
+									"    pm.expect(rate.dislikes.toString()).equal(target.dislikes.toString(), 'Количество дизлайков должено соответствовать количеству дизлайков события с указанным идентификатором');\r",
 									"});"
 								],
 								"type": "text/javascript"
@@ -620,7 +624,9 @@
 									"pm.expect(target).to.have.property('createdOn');\r",
 									"pm.expect(target).to.have.property('location');\r",
 									"pm.expect(target).to.have.property('requestModeration');\r",
-									"pm.expect(target).to.have.property('rate');\r",
+									"pm.expect(target).to.have.property('rating');\r",
+									"pm.expect(target).to.have.property('likes');\r",
+									"pm.expect(target).to.have.property('dislikes');\r",
 									"});\r",
 									"\r",
 									"pm.test(\"Данные в ответе должны соответствовать данным в запросе\", function () {\r",
@@ -2378,11 +2384,14 @@
 									"const target = pm.response.json();\r",
 									"const isSorted = require('lodash');\r",
 									"\r",
-									"pm.test(\"Пользователи должны содержать поля: id, name, email, rate\", function () {\r",
+									"pm.test(\"Пользователи должны содержать поля: id, name, email, rating, likes, dislikes, privateRating\", function () {\r",
 									"    pm.expect(target[0]).to.have.property('id');\r",
 									"    pm.expect(target[0]).to.have.property('name');\r",
 									"    pm.expect(target[0]).to.have.property('email');\r",
-									"    pm.expect(target[0]).to.have.property('rate');\r",
+									"    pm.expect(target[0]).to.have.property('rating');\r",
+									"    pm.expect(target[0]).to.have.property('likes');\r",
+									"    pm.expect(target[0]).to.have.property('dislikes');\r",
+									"    pm.expect(target[0]).to.have.property('privateRating');\r",
 									"});\r",
 									"\r",
 									"pm.test(\"Должен быть найден только 4-е пользователя по заданному фильтру\", function () {\r",
@@ -2390,7 +2399,7 @@
 									"});\r",
 									"\r",
 									"pm.test(\"Данные в ответе должны быть отсортированы по публичному рейтингу\", function () {\r",
-									"    var expectedSortedOrder = isSorted.orderBy(target, ['rate.rating'],['desc']);\r",
+									"    var expectedSortedOrder = isSorted.orderBy(target, ['rating'],['desc']);\r",
 									"\r",
 									"    pm.expect(target).to.eql(expectedSortedOrder);\r",
 									"});"
@@ -2654,12 +2663,14 @@
 									"const source = pm.collectionVariables.get('response');\r",
 									"const target = pm.response.json();\r",
 									"\r",
-									"pm.test(\"Подборка должны содержать поля: id, title, pinned, events, rate\", function () {\r",
+									"pm.test(\"Подборка должны содержать поля: id, title, pinned, events, rating, likes, dislikes\", function () {\r",
 									"pm.expect(target).to.have.property('id');\r",
 									"pm.expect(target).to.have.property('title');\r",
 									"pm.expect(target).to.have.property('pinned');\r",
 									"pm.expect(target).to.have.property('events');\r",
-									"pm.expect(target).to.have.property('rate');\r",
+									"pm.expect(target).to.have.property('rating');\r",
+									"pm.expect(target).to.have.property('likes');\r",
+									"pm.expect(target).to.have.property('dislikes');\r",
 									"});\r",
 									"\r",
 									"pm.test(\"Данные в ответе должны соответствовать данным в запросе\", function () {\r",
@@ -2762,16 +2773,18 @@
 									"const isSorted = require('lodash');\r",
 									"target.forEach(function(element){if (element.id == source.id) founded = element});\r",
 									"\r",
-									"pm.test(\"Подборка должны содержать поля: id, title, pinned, events, rate\", function () {\r",
+									"pm.test(\"Подборка должны содержать поля: id, title, pinned, events, rating, likes, dislikes\", function () {\r",
 									"pm.expect(target[0]).to.have.property('id');\r",
 									"pm.expect(target[0]).to.have.property('title');\r",
 									"pm.expect(target[0]).to.have.property('pinned');\r",
 									"pm.expect(target[0]).to.have.property('events');\r",
-									"pm.expect(target[0]).to.have.property('rate');\r",
+									"pm.expect(target[0]).to.have.property('rating');\r",
+									"pm.expect(target[0]).to.have.property('likes');\r",
+									"pm.expect(target[0]).to.have.property('dislikes');\r",
 									"});\r",
 									"\r",
 									"pm.test(\"Данные в ответе должны быть отсортированы по публичному рейтингу\", function () {\r",
-									"    var expectedSortedOrder = isSorted.orderBy(target, ['rate.rating'],['desc']);\r",
+									"    var expectedSortedOrder = isSorted.orderBy(target, ['rating'],['desc']);\r",
 									"\r",
 									"    pm.expect(target).to.eql(expectedSortedOrder);\r",
 									"});\r",
